Move feed notification service layer
