+++ services/feed/action_test.go
@@ -1,7 +1,7 @@
 // Copyright 2019 The Gitea Authors. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-package action
+package feed
 
 import (
 	"path/filepath"
